requirements.txt
langchain
langchain-core
langchain-google-genai
pandas
google-generativeai
python-dotenv
pydantic